refactor: extract shell layout from App
